fix: Allow pull datasets without records (#3851)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

Fixed for `ds.pull()` with datasets without records.

Also, the `__len__` for the filtered dataset returns a 0 instead of
`None`. Otherwise, the `len(ds)` fails with a confusing error to the
user.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Alvaro Bartolome <alvaro@argilla.io>