feat: add records metadata validation on the backend (#3862)
# Description

This PR adds a new column `allow_extra_metadata` to the `datasets`
table, that allows defining if the records of a dataset can contain
extra metadata (i.e. there is no metadata property created with the same
name). In addition, this PR adds the required logic to validate that the
provided metadata is valid based on the metadata properties created for
a dataset and its `allow_extra_metadata`.

**Type of change**

- [x] New feature (non-breaking change which adds functionality)

**How Has This Been Tested**


**Checklist**

- [ ] I added relevant documentation
- [x] I followed the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the `CHANGELOG.md` file (See
https://keepachangelog.com/)