fix: `sort_by` by metadata property validation (#3885)
# Description

This PR fixes the expected value that has to be provided to the
`sort_by` query param of the listing records endpoints when using a
metadata property. Before, it was just expecting the name of the
metadata property `sort_by=metadata-property-name:asc`, which is not
correct. Now, it expects `sort_by=metadata.metadata-propert-name:asc`.

**Type of change**

- [x] Bug fix (non-breaking change which fixes an issue)

**How Has This Been Tested**


**Checklist**

- [x] I followed the style guidelines of this project
- [x] I did a self-review of my code
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the `CHANGELOG.md` file (See
https://keepachangelog.com/)