chore: upgrade `setfit` and remove `huggingface_hub` restriction (#3939)
# Description

This PR upgrades `setfit` to 0.7.0 or higher, as there was an
incompatibility issue with `huggingface_hub` 0.12 or higher, which was
preventing us to install the latest `huggingface_hub` version within the
`integrations` extra requirements. So on, this PR akso removes the
`huggingface_hub` restriction (while keeping the previous one `>=0.5.0`
for backwards compatibility).

From now on, everyone should be able to `pip install
argilla[integrations]` without a dependency version mismatch issue that
was being triggered before due to the incompatibility of `datasets` and
`huggingface_hub`, which was unrelated to `setfit`, but since the
previous restriction on `huggingface_hub` was due to `setfit`, patching
`setfit` fixes the rest of the issues.

**Type of change**

- [x] Dependency upgrade / fix (upgrades/downgrades/fixes a dependency
related issue)

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [x] Run `pip install -e ".[integrations]"` with a Python matrix from
3.8 to 3.11 and then running `pytest -s tests/integration` to re-run the
integration tests to ensure that the fix does not affect any

**Checklist**

- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)