refactor: Using the `metadata.name` instead `metadata.id` for search index (#3994)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

This PR changes how metadata info is indexed in Elasticsearch, by using
names instead of ids. The aim of this code refactoring is to make raw
indexes more user-friendly and let the possibility of using the Lucene
query DSL for filtering data from Argilla (as we already do for other
tasks).

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>