fix: Support combining knn and filtering with OpenSearch  (#4111)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

For OpenSearch engines, including boolean filters as part of the `knn`
filter may return query phase errors if the response fields are not
found in documents for the requested user (`responses.username.status =
...`). This is something that is not happening with Elasticsearch.

So, this PR fixes this by setting those query filters as a `post_filter`
in the query. (This will potentially return less record than expected
`k`, but it will be something to assume when using the OpenSearch
engine.

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

Some acceptance tests have been launched locally. 

**Checklist**

- [ ] I followed the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the `CHANGELOG.md` file (See
https://keepachangelog.com/)