feat: add search records suggestions options new endpoint (#4260)
# Description

This PR adds a new `GET
/datasets/:dataset_id/records/search/suggestions/options` endpoint that
will return information about the options that are allowed to search
records with suggestions.

Right now we are returning the list of different agents that are defined
for a question suggestions inside the dataset.

The response is in the form:

```json
{
    "items": [
        {
            "question": {
                "id": "5d87c0e7-9eaf-4b92-ba0d-6cf50a4176e6",
                "name": "question-a",
            },
            "agents": [],
        },
        {
            "question": {
                "id": "bd68eec1-cace-4df8-b239-9da081ac41b3",
                "name": "question-b",
            },
            "agents": [
                "agent-a"
            ],
        },
        {
            "question": {
                "id": "30ccbe15-8e9a-4893-8fce-e713a1c32ea1",
                "name": "question-c",
            },
            "agents": [
                "agent-a",
                "agent-b",
                "agent-c"
            ],
        },
    ]
}
```

Details about the endpoint behavior and code changes:
* All the questions of the dataset will have an entry in the response,
event those that doesn't have questions or have questions but without
suggestions.
* In the list of `agents` we will only include different agent strings
so you should not see any repetition.
* We have included two implementations of the function that return the
results one for SQLite and another one for PostgreSQL, this is related
with the lack of support from SQLite to arrays and the impossibility of
aggregating elements as arrays with this database.
* The SQLite implementation is using `group_concat` specifically, this
is doing a string concatenation using a `,` character as delimiter. This
is open to errors if the agent is using that character inside the agent
string. We will add a validation for this on issue #4237 to avoid
possible errors.
* We have added a new `server/contexts/search.py` file, where we want to
add domain functions associated with the search functionality or
argilla.

Closes #4224 

**Type of change**

- [x] New feature (non-breaking change which adds functionality)

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [x] Running locally unit tests using SQLite database.
- [x] Running locally unit tests using PostgreSQL database.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)

---------

Co-authored-by: Gabriel Martín Blázquez <gmartinbdev@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>