feat: add validations for records search filters and sort (#4306)
# Description

This PR add some changes related with how we search for records:
* Add a new `SearchRecordsQueryValidator` class that allow us to
validate some details about a schema `SearchRecordsQuery`.
* Integrate this validation class into handlers doing searching for
records. Returning a `422` error if this validation is not successful.
* Changed `gte` and `lte` attributes on `RangeFilter` to `ge` and `le`.
The change is inspired on a previous use of `ge` and `le` for a Metadata
schema so we don't want to add a new way to specify these kind of
ranges.
* Add a new `RecordFilterScope` to be used on the query `sort` attribute
like this:
```json
{
  "sort": [
    {"scope": {"entity": "record", "property": "inserted_at"}, "order": "desc"},
    {"scope": {"entity": "record", "property": "updated_at"}, "order": "desc"},
  ]
}
```
* I didn't add any validation to avoid `RecordFilterScope` to be used on
`filters` attribute too so it should be ignored on this section for the
search engine /cc @frascuchon
* Some more small changes, for example the addition of
`get_question_by_name_and_dataset_id_or_raise` and
`get_metadata_property_by_name_and_dataset_id_or_raise` to be used on
`SearchRecordsQueryValidator` but that could be useful also once we fix
current exception handling with FastAPI.

Refs #4227 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

- [x] Running unit tests locally.

**Checklist**

- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)