feat: index records with suggestions for search engine (#4317)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

This PR adds support for indexing suggestions when indexing records in
the search index.

In order to be sure that all record attributes are passed when indexing
records, a workaround has been implemented by forcing a record,
responses, and suggestions to refresh before indexing them. Otherwise is
hard to add that info when loading records in the current code base.

This behavior must be reviewed and simplified cc @gabrielmbmb @jfcalvo 

Changes related to tests will be moved into a separate PR since extra
refactoring work must be done. Once this extra PR is merged here, the PR
will be marked as ready for review. Related PR
https://github.com/argilla-io/argilla/pull/4318


Refs: #3849 

Closes https://github.com/argilla-io/argilla/issues/4230

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: José Francisco Calvo <jose@argilla.io>
Co-authored-by: José Francisco Calvo <josefranciscocalvo@gmail.com>