perf: improve upsert responses bulk (#4451)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

This PR applies some performance improvements related to the bulk upsert
responses endpoint since some environments are experimenting with a high
execution time when running this operation with 100 items in the
request.

The performance improvements are:

1. Read all requested records at once instead of once at a time.
2. Stop checking the existence of the search index since is taking
longer.

These improvements are applied based on the profiling results using the
test ...

Also, since the current API controller is doing many things, a new
design proposal is applied in the implementation by defining a simple
use case implementation for this functionality. The use case wraps all
the needed steps to cover the functionality and make the controller
implementation quite simple, focusing only on exception handling and
request/response serialization.

Closes #4449 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] New feature (non-breaking change which adds functionality)
- [X] Refactor (change restructuring the codebase without changing
functionality)
- [X] Improvement (change adding some improvement to an existing
functionality)

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

Test locally

**Checklist**

- [ ] I added relevant documentation
- [X] I followed the style guidelines of this project
- [X] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [X] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the `CHANGELOG.md` file (See
https://keepachangelog.com/)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>